Fixed bug in counting hardness level for matches (added field for place in Team)

+++ src/main/java/com/jamex/refereestaffer/model/dto/TeamDto.java
@@ -18,5 +18,5 @@ public class TeamDto {
     @NotNull
     private final String city;
 
-    private final Integer points;
+    private final Short points;
 }
